Adding README with view-indpendent build and test instructions
Summary:
Adding a README to ipu_tensorflow_addons with build and test instructions which don't rely on the scripts in tensorflowview.
As part of the review it would be a good idea to try following the instructions (See the Test Plan).

Note: We've avoided mentioning the specific TF version (e.g. 2.4.4) anywhere as we would then have to remember to maintain the number. For now people will only get this shipped with the poplar sdk anyway so I think it is a future issue.

This is identical to the v2 diff, other than the bazel version and the example test target (both in README.md).

TF1.15 Only

V2 diff: D55722

Test Plan:
Before landing this someone should independently go through the instructions and try to do the following:
 - Build a wheel file.
 - Run all the tests and have them pass.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgew, jackh, markf

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jackh

Subscribers: georgep, georgew, jackh, markf

Maniphest Tasks: T49706

Differential Revision: https://phabricator.sourcevertex.net/D55970