Handle creation of slot variables correctly in optimizers
Summary:
By default when adding a slot variable (for example when restoring from a checkpoint), it matches the dtype of the variable.
This resolves this by overriding add_slot which creates the variable with a right dtype.
Fix T53780

TF2.5 Only

Test Plan: Adapted the tests to save/load weights which would have failed before.

Reviewers: adams, jackh, samuelh, #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: samuelh, #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T53780

Differential Revision: https://phabricator.sourcevertex.net/D62305