Get carry-over tensor size for each layer in a keras model
Summary:
For a layer, a carry-over tensor is a tensor not used by the layer but is an input argument to later layers.
In this diff I implemented a function to find size of such tensor for every layer.
Plus carry-over + input arg (If a tensor is both carry-over and input, it is counted only once)
Plus carry-over + output
Plus a few other helper functions

Test Plan:
A test file has been added. The test passes locally.
The test has not been added to CI because some dependencies are missing for this diff.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jackh

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jackh

Maniphest Tasks: T64995

Differential Revision: https://phabricator.sourcevertex.net/D73029